requirements.txt
streamlit==1.38.0
pymupdf
aiohttp
google-generativeai
